Documentando adaptador
Moviendo el adaptador genérico a su propio archivo
Comenzando documentación en yard
